BACKPORT: arm64: tlbflush: remove redundant ASID casts to (unsigned long)
The ASID macro returns a 64-bit (long long) value, so there is no need
to cast to (unsigned long) before shifting prior to a TLBI operation.


Bug: 31432001
(cherry picked from commit f3e002c24e1f3b66f6e392ecd6928b5d04672c54)

+++ arch/arm64/include/asm/tlbflush.h
@@ -111,7 +111,7 @@ static inline void flush_tlb_mm(struct mm_struct *mm)
 	dsb();
 	isb();
 #else
-	unsigned long asid = (unsigned long)ASID(mm) << 48;
+	unsigned long asid = ASID(mm) << 48;
 
 	dsb(ishst);
 	__tlbi(aside1is, asid);
@@ -129,8 +129,7 @@ static inline void flush_tlb_page(struct vm_area_struct *vma,
 	dsb();
 	isb();
 #else
-	unsigned long addr = uaddr >> 12 |
-		((unsigned long)ASID(vma->vm_mm) << 48);
+	unsigned long addr = uaddr >> 12 | (ASID(vma->vm_mm) << 48);
 
 	dsb(ishst);
 	__tlbi(vale1is, addr);
